BIB-54

+++ src/main/java/com/svydovets/bibirnate/session/impl/SessionImpl.java
@@ -14,7 +14,6 @@ import com.svydovets.bibirnate.cache.CacheUtils;
 import com.svydovets.bibirnate.exceptions.BibernateException;
 import com.svydovets.bibirnate.exceptions.JdbcException;
 import com.svydovets.bibirnate.exceptions.PersistenceException;
-import com.svydovets.bibirnate.exceptions.TransactionManagerException;
 import com.svydovets.bibirnate.jdbc.JdbcEntityDao;
 import com.svydovets.bibirnate.logs.SqlLogger;
 import com.svydovets.bibirnate.session.Session;
@@ -72,27 +71,23 @@ public class SessionImpl implements Session {
      */
     @Override
     public void remove(Object entity) {
+        Objects.requireNonNull(entity);
         log.trace("Removing {} by id", entity.getClass().getSimpleName());
         checkIfSessionClosed();
         var id = EntityUtils.getEntityIdValue(entity);
         var result = CacheUtils.extract(cacheContainer, entity.getClass(), id);
-//        if (result.isPresent()) {
+        if (result.isPresent()) {
             try {
-                try {
-                    transactionManager.begin();
-                    jdbcEntityDao.remove(entity);
-                    CacheUtils.invalidate(cacheContainer, entity);
-                    transactionManager.commit();
-                } catch (Exception ex) {
-                    transactionManager.rollback();
-                }
-            } catch (TransactionManagerException ex) {
-                throw new PersistenceException("Could not rollback transaction", ex);
+                connection.setAutoCommit(false);
+                jdbcEntityDao.remove(entity);
+                CacheUtils.invalidate(cacheContainer, entity);
+            } catch (SQLException ex) {
+                throw new RuntimeException(ex);
             }
-//        } else {
-//            throw new PersistenceException(
-//              String.format("Object %s can not be deleted as it is not present in Session", entity));
-//        }
+        } else {
+            throw new PersistenceException(
+              String.format("Object %s can not be deleted as it is not present in Session", entity));
+        }
     }
 
     /**
